+++ src/storage-adapter-s3/env/index.js
@@ -24,6 +24,7 @@ const getCredentials = async provider => {
     if (provider === 'aws') prefix = 'aws'
     if (provider === 'cloudflare') prefix = 'cf'
     if (provider === 'hetzner') prefix = 'hetzner'
+    if (provider === 'backblaze') prefix = 'backblaze'
     credentials = {
       [process.env[`${prefix.toUpperCase()}_PROFILE`]]: {
         [`${prefix}_account_id`]:
@@ -103,6 +104,28 @@ const hetzner = async (cfg, ctx) => {
   return cfg
 }
 
+const backblaze = async (cfg, ctx) => {
+  if (!ctx.specifiedTypes.includes('s3')) {
+    return cfg
+  }
+
+  if (!cfg.backblaze) {
+    return cfg
+  }
+
+  let credentials
+  credentials = await getCredentials('backblaze')
+  credentials = credentials[process.env.BACKBLAZE_PROFILE]
+
+  cfg.backblaze.storageClient.endpoint = `https://s3.${cfg.backblaze.storageClient.region}.backblazeb2.com`
+  cfg.backblaze.storageClient.credentials = {
+    accessKeyId: credentials.backblaze_access_key_id,
+    secretAccessKey: credentials.backblaze_secret_access_key
+  }
+
+  return cfg
+}
+
 const loadConfig = async (environment = process.env.NODE_ENV) => {
   let ctx = {}
   ctx.specifiedTypes = process.argv.filter(p => p.startsWith('--type')).map(p => p.split('=')[1])
@@ -125,6 +148,7 @@ const loadConfig = async (environment = process.env.NODE_ENV) => {
   cfg = await aws(cfg, ctx)
   cfg = await cloudflare(cfg, ctx)
   cfg = await hetzner(cfg, ctx)
+  cfg = await backblaze(cfg, ctx)
 
   return cfg
 }